+++ NEWS.md
@@ -1,8 +1,28 @@
 # anvil (development version)
 
+## New Features
+
+* New primitives:
+  * `nvl_cholesky()` to compute the Cholesky decomposition of a matrix.
+  * `nvl_triangular_solve()` to solve a system of linear equations with a triangular matrix.
+* New API functions:
+  * `nv_diag()` to create a diagonal matrix from a 1-D tensor.
+  * `nv_eye()` to create an identity matrix.
+  * `nv_solve()` to solve a system of linear equations.
+  * `nv_cholesky()` to compute the Cholesky decomposition of a matrix.
+* Improved semantics:
+  * `nvl_cholesky()` now zeros out the upper/lower triangle of the output.
+
 ## Bug Fixes
 
 * +-Inf/NaN are correctly created for `f64` when inlined into the XLA exectuable (#182).
+  This caused wrong results with e.g. `nv_reduce_max()` when working with `f64`.
+* Corrected argument checks in `nv_iota()`.
+
+## Documentation
+
+* New vignette on implementing Gaussian Processes.
+* New vignette on implementing Metropolis-Hastings sampling.
 
 # anvil 0.1.0